split url upload form into its own component
